BLP Estimation

Estimates the best linear predictor (BLP) of the actual CATEs using the estimated CATEs.

blp_estimation(
  Y,
  D,
  cates,
  pscore,
  mu,
  mu0,
  mu1,
  scores,
  strategies,
  denoising
)

Arguments

Y

Observed outcomes.

D

Treatment indicator.

cates

CATE predictions. Must be produced by a model estimated using different observations than those in Y and D.

pscore

Propensity scores predictions. Must be produced by a model estimated using different observations than those in Y and D (unless the propensity score is known, in which case we provide the true values).

mu

Conditional mean predictions. Must be produced by a model estimated using different observations than those in Y and D

mu0

Control units' conditional mean predictions. Must be produced by a model estimated using different observations than those in Y and D

mu1

Treated units' conditional mean predictions. Must be produced by a model estimated using different observations than those in Y and D

scores

Estimated doubly-robust scores. Must be estimated via K-fold cross-fitting using the same observations as in Y and D.

strategies

Character vector controlling the identification strategies to implement for BLP and GATES. Admitted values are "WR" (weighted residuals), "HT" (Horwitz-Thompson), and "AIPW" (augmented inverse-probability weighting).

denoising

Character vector controlling if and which additional covariates to include in the regressions to reduce the variance of the estimation. Admitted values are "none", "cddf1", "cddf2", "mck1", "mck2", and "mck3".

Value

A list of fitted models as lm_robust objects.

Details

To estimate the BLP of the actual CATEs using the estimated CATEs, the user must provide observations on the outcomes and the treatment status of units in the validation sample, as well as their estimated cates and nuisance functions. Be careful, as these estimates must be obtained using only observations from the training sample (see the example section below). Additionally, the user must provide doubly-robust scores estimated in the validation sample using K-fold cross fitting.

The BLP is estimated using three different strategies, all involving fitting suitable linear models. For each of these strategis, different model specifications are considered that differ in additional and optional covariates that can be included in the regressions to reduce the estimation variance. Check the online short tutorial and denoising vignette for details.

Standard errors are estimated using the Eicker-Huber-White estimator.

See also

Author

Riccardo Di Francesco